put all magic numbers in constants, changed return type of getArea() method to double

+++ src/main/java/core/basesyntax/Circle.java
@@ -18,7 +18,7 @@ public class Circle extends Figure {
     }
 
     @Override
-    public int getArea() {
-        return (int) (Math.PI * radius * radius);
+    public double getArea() {
+        return (Math.PI * radius * radius);
     }
 }
